suppression des pages inutiles et double to int

+++ server/play-samples-play-scala-hello-world-tutorial/app/controllers/Drone.scala
@@ -1,6 +1,6 @@
 package controllers
 
-case class Drone(id: Double, latitude: String, longitude: String, temperature: Double, battery: Double, weight: Double)
+case class Drone(id: Int, latitude: String, longitude: String, temperature: Int, battery: Int, weight: Int)
 
 object Drone {
   import play.api.libs.json._
@@ -8,12 +8,12 @@ object Drone {
 
     // convert from JSON string to a Stock object (de-serializing from JSON)
     def reads(json: JsValue): JsResult[Drone] = {
-      val id = (json \ "id").as[Double]
+      val id = (json \ "id").as[Int]
       val latitude = (json \ "latitude").as[String]
       val longitude = (json \ "longitude").as[String]
-      val temperature = (json \ "temperature").as[Double]
-      val battery = (json \ "battery").as[Double]
-      val weight = (json \ "weight").as[Double]
+      val temperature = (json \ "temperature").as[Int]
+      val battery = (json \ "battery").as[Int]
+      val weight = (json \ "weight").as[Int]
       JsSuccess(Drone(id, latitude, longitude, temperature, battery, weight))
     }
